Source Changes
I'm trying to create a test branch and a main branch, both of which can be bugtested on 5e.tools simultaneously. However, I want to make sure there's a difference between the two, so I'm altering the sources to make sure I can tell which "item" or "spell" is from which compendium. As such, I'll have to change each source by hand before committing? Is there a point to that? idk...